Release v1.0.0

+++ components/clarinet-cli/src/frontend/telemetry.rs
@@ -5,13 +5,13 @@ use segment::{
     Client, HttpClient,
 };
 
-use orchestra_types::StacksNetwork;
+use chainhook_types::StacksNetwork;
 
 pub enum DeveloperUsageEvent {
     NewProject(DeveloperUsageDigest),
     PokeExecuted(DeveloperUsageDigest),
     CheckExecuted(DeveloperUsageDigest),
-    TestSuiteExecuted(DeveloperUsageDigest, bool, u32),
+    TestSuiteExecuted(DeveloperUsageDigest, bool, usize),
     DevnetExecuted(DeveloperUsageDigest),
     ProtocolPublished(DeveloperUsageDigest, StacksNetwork),
     DebugStarted(DeveloperUsageDigest, u32),
